updated file structure, added bad input check and stock_lookup tests
