+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "python.pythonPath": "home/emonnier/.pyenv/versions/3.8.2/bin/python"
+  "python.pythonPath": ".venv/bin/python"
 }
