+++ pyproject.toml
@@ -14,8 +14,9 @@ aylien-apiclient = "^0.7.0"
 python-dotenv = "^0.13.0"
 alive-progress = "^1.5.1"
 
-
 [tool.poetry.dev-dependencies]
+pytest = "^5.4.3"
+mock = "^4.0.2"
 
 [build-system]
 requires = ["poetry>=0.12"]
