--- settings.py
@@ -1,3 +1,0 @@
-# settings.py
-from dotenv import load_dotenv
-load_dotenv()
